Implement ReplayParked and RestartSubsystems
I'm skipping adding tests for RestartSubsystems since I suspect that
it's a bit destructive on the server-side and may cause some pauses or
instability in test cases executed after it is restarted. This might
be revisited later.

I'm also skipping implementing the new List RPC since we have an
acceptable albeit hacky workaround that has better server version
coverage. This can be revisited later if we decide to check the
server capabilities support.