Check HTTP/2 window size for acks/nacks and batch appends
When sending regular requests like read and append on the connection,
we check that the HTTP/2 windows have enough bytes to fit the requests.
The handle_cast implementation used for sending acks, nacks and batch
appends did not check this though. On a busy connection this could
result in dropped messages.

This change refactors acks, nacks and batch appends to stream in the
same way as regular requests. This requires some changes in
Spear.Connection.Request to allow a request's 'continuation' to be
complete (finished or halted) and still send any remaining buffered
data.