Remove EnableExternalTcp from eventstore.conf
This is an unknown option in the latest versions of EventStore and
that causes an error on startup. We can suppress that error with the
new AllowUnknownOptions option, but that option isn't known by older
EventStore versions, so they error out because the option doesn't exist,
failing the compatibility CI.

So we'll just remove it.

    (ノಠ益ಠ)ノ彡┻━┻